pyproject.toml
[tool.poetry]
name = "sklearn-ann"
version = "0.1.0"
description = "Various integrations for ANN (Approximate Nearest Neighbours) libraries into scikit-learn."
authors = ["Frankie Robertson <frankie@robertson.name>"]
license = "BSD"

[tool.poetry.dependencies]
python = "^3.8"
scikit-learn = "^0.24.0"
pytest = {version = "^6.2.1", optional = true}
pytest-cov = {version = "^2.10.1", optional = true}
Sphinx = {version = "^3.4.1", optional = true}
sphinx-gallery = {version = "^0.8.2", optional = true}
sphinx-rtd-theme = {version = "^0.5.0", optional = true}
numpydoc = {version = "^1.1.0", optional = true}
matplotlib = {version = "^3.3.3", optional = true}
annoy = {version = "^1.17.0", optional = true}
faiss-cpu = {version = "^1.6.5", optional = true}
flake8 = {version = "^3.8.4", optional = true}
black = {version = "=19.10b0", optional = true}
pynndescent = {version = "^0.5.1", optional = true}
pandas = {version = "^1.2.0", optional = true}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.extras]
tests = ["pytest", "pytest-cov", "flake8", "black"]
docs = ["sphinx", "sphinx-gallery", "sphinx_rtd_theme", "numpydoc", "matplotlib"]
annoy = ["annoy"]
faiss = ["faiss-cpu"]
pynndescent = ["pynndescent"]
